perf: stop finalizing packets and building per-block ones nothing fills
Follow-up to #321, which left the WorldPacket object as the largest single
allocation in the Alterac Valley trace.

ByteBuffer had the only finalizer in the codebase. A finalizable type cannot
use the runtime's fast allocation path and is registered with the finalization
queue as it is built, which measured ~4x the cost of the object itself and was
paid by every packet in both directions — while every packet is already
disposed. A rental that does miss its Dispose is now collected like any other
array, which is also the safer failure: the finalizer could hand a buffer back
to the pool while a read-mode GetData() caller still held it. Dropping the
unused compat stream and narrowing the receive time to the int it always held
takes the object from 64 to 48 bytes.

SpellStart and SpellGo each built a SpellCastData — five lists, a target and a
heal prediction — that the handler overwrote on the next line.

The AuraUpdate and PowerUpdate built for every Values and create block are now
reused until a block fills one, at which point it is handed on and the next
block builds a fresh one. On a 3.3.5a backend almost no block fills either:
auras arrive through SMSG_AURA_UPDATE, since the legacy field block has no
UNIT_FIELD_AURA at all, and powers are only forwarded for the player and their
pet. The scratch is taken at the point the constructor used to run, so
PowerUpdate keeps carrying the pre-correction guid exactly as before.

UnitData.NpcFlags moves into the object as an inline array. AzerothCore sends
UNIT_NPC_FLAGS in every unit and player Values block, so the lazy array was
allocated for nearly all of them.

NPCFlagsVanilla and PlayerFlagsLegacy become uint-backed. As ints, a legacy
value with bit 31 set is negative, and the checked conversion in CastFlags
threw OverflowException out of the whole update packet instead of translating
the bits the modern enum does have.

Measured on the M4 with --job medium, against master:

  ReceiveLegacyPacket                 36.31 ns / 64 B -> 17.62 ns / 48 B
  SendLegacyPacket                    31.26 ns / 64 B -> 10.25 ns / 48 B
  SendModernPacket (ByteBuffer path)  63.22 ns / 256 B -> 41.84 ns / 240 B
  UpdateObject_ValuesBatch            16,334 ns / 6,983 B -> 15,390 ns / 4,928 B
  SpellStart                          323.5 ns / 1,178 B -> 243.3 ns / 648 B
  SpellGo                             355.0 ns / 1,270 B -> 274.6 ns / 744 B

The Gen1 collections the spell arms reported before are gone; that was
finalization promoting packets out of gen0.

An 18-minute AzerothCore playerbots session agrees: SMSG_SPELL_GO 1,303 ->
749 B and SMSG_SPELL_START 1,136 -> 649 B per packet, gen0 6 -> 4 per minute,
0.36 -> 0.30 MB/s allocated, and no errors in the log.

The benchmarks run the real handlers through a new harness in
HermesProxy.Tests/Support, compile-linked into the benchmark project so a
benchmark and its correctness test drive the same packets. The wire tests pin
the bytes captured from the handlers before any of this.

+++ HermesProxy/World/Packet.cs
@@ -21,7 +21,6 @@ using Framework.Logging;
 using System;
 using System.Buffers;
 using System.Buffers.Binary;
-using HermesProxy.Enums;
 using HermesProxy.World.Enums;
 using HermesProxy.World.Client;
 using HermesProxy.World.Logging;
@@ -399,11 +398,13 @@ public class WorldPacket : ByteBuffer
             return LegacyVersion.GetUniversalOpcode(GetOpcode());
     }
 
-    public long GetReceivedTime() { return m_receivedTime; }
-    public void SetReceiveTime(long receivedTime) { m_receivedTime = receivedTime; }
+    public int GetReceivedTime() { return m_receivedTime; }
+    public void SetReceiveTime(int receivedTime) { m_receivedTime = receivedTime; }
 
     uint opcode;
-    long m_receivedTime;
+    // An Environment.TickCount, which is an int and wraps every ~25 days. Held as a long, it cost
+    // this object eight bytes (and, with the alignment, sixteen) for no extra range.
+    int m_receivedTime;
 }
 
 /// <summary>
